Fix leaks in Azure Blob parallelism

+++ core/include/storage_manager/storage_azure_blob.h
@@ -136,7 +136,7 @@ class AzureBlob : public StorageCloudFS {
 
   size_t max_stream_size_ = 1024;
 
-  size_t num_threads_ = std::thread::hardware_concurrency()/2;
+  int num_threads_ = std::thread::hardware_concurrency()/2;
 
   std::string get_path(const std::string& path);
 
@@ -178,7 +178,7 @@ class AzureBlob : public StorageCloudFS {
   std::future<storage_outcome<void>> upload_block_blob(const std::string &blob, uint64_t block_size, int num_blocks,
                                                        std::vector<std::string> block_list,
                                                        const char* buffer, uint64_t bufferlen,
-                                                       uint parallelism=1);
+                                                       int parallelism=1);
 
   void update_expected_filesizes_map(const std::string& path, size_t size) {
     auto search = filesizes_map_.find(path);
